All constants have been moved to RustConstants
RustConstants now consists of RustColorConstants, RustConstants,
RustPreferenceConstants and RustTextConstants. All reference to those
classes have been updated to RustConstants.
