--- rustycage-plugin/src/org/rustycage/preferences/PreferenceConstants.scala
@@ -1,9 +1,0 @@
-package org.rustycage.preferences
-
-object PreferenceConstants {
-  final val SAVE_BEFORE_COMPILE: String = "saveBeforeCompile"
-  final val RUST_C: String = "pathRustCompiler"
-  final val P_PATH: String = "pathPreference"
-  final val P_BOOLEAN: String = "booleanPreference"
-  final val P_CHOICE: String = "choicePreference"
-}
