+++ rustycage-plugin/src/org/rustycage/preferences/RustPreferencePage.scala
@@ -1,6 +1,7 @@
 package org.rustycage.preferences
 
 import org.rustycage.RustPlugin
+import org.rustycage.PreferenceConstants
 import org.eclipse.jface.preference.{BooleanFieldEditor, DirectoryFieldEditor, FieldEditorPreferencePage}
 import org.eclipse.ui.{IWorkbench, IWorkbenchPreferencePage}
 
@@ -15,6 +16,7 @@ class RustPreferencePage extends FieldEditorPreferencePage(FieldEditorPreference
     addField(new DirectoryFieldEditor(RUST_C, "&Rust compiler:", getFieldEditorParent))
     addField(new DirectoryFieldEditor(P_PATH, "Rust home:", getFieldEditorParent))
     addField(new BooleanFieldEditor(SAVE_BEFORE_COMPILE, "Save before compile", getFieldEditorParent))
+    addField(new DirectoryFieldEditor(CARGO_LOCATION, "Cargo location:", getFieldEditorParent))
   }
 
   override def init(workbench: IWorkbench) {
